requirements.txt
groq
streamlit
sentence-transformers
numpy
scipy
nltk
pdfplumber
PyPDF2
python-dotenv
